Create Exercise 1
Problem:
Create a computer program that derives the average from N random numbers obtained from a uniformly random distribution of numbers in space [0 to 1]. The program must run for N = 10, 100, 1000, 10000, 100000, 1000000 random numbers. Make the graph of the mean as a function of N (where the N axis is logarithmic). Describe what conclusions you draw from the results. Use your personal ID number as the original seed (as in all subsequent problems).